% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{aggregate_csvs}
\alias{aggregate_csvs}
\title{combine data csvs}
\usage{
aggregate_csvs(csv_list, spec.df)
}
\arguments{
\item{csv_list}{list of files,e.g.}
}
\description{
given a list of CSV files, read and check the specs,
then aggregate into a single DF.
}
