% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.Date.flexible}
\alias{as.Date.flexible}
\title{flexible character to date converter}
\usage{
\method{as.Date}{flexible}(x, ...)
}
\arguments{
\item{x}{character to be converted to date, or empty string}
}
\value{
Date or NA if x is blank, using as.Date function with 'tryFormats' option
}
\description{
try to convert a character to date using one of two delimiters ('/' or '-') and
start with international data format first (d/m/Y), but if that fails try sci date format (Y-m-d)
and finally US format.  Value can be 'optional' in that if it's blank or not a date, NA is return
}
