% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\docType{data}
\name{commecology_rules_biomass_str}
\alias{commecology_rules_biomass_str}
\title{Column Definitions for CommecologyRULEs biomass datasets}
\format{
\subsection{\code{commecology_rules_biomass_str}}{

A data frame 40-ish rows and 4 columns:
\describe{
\item{col_name}{name of column in biomass}       ""
\item{col_description}{desc}
\item{data_str}{R data type}
\item{example}{what may go in the data}
}
}
}
\usage{
commecology_rules_biomass_str
}
\description{
list of column names with meanings, data type, units and examples
}
\keyword{datasets}
