% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gdrive_client_setup}
\alias{gdrive_client_setup}
\title{get a google drive 'client' for authentication from env file}
\usage{
gdrive_client_setup()
}
\value{
'client' for use in drive_auth_configure or
}
\description{
reads values from the enviroment for configurating a google drive client
for accesing gdrive file or gsheets data
}
