% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gdrive_setup}
\alias{gdrive_setup}
\title{connect to your google drive account, required set-up for using the google drive packages}
\usage{
gdrive_setup(drive_email = NULL)
}
\value{
TRUE if the functions requiring the google R packages complete without error
}
\description{
this is wrapped in a function because it has the side effect of logging in, and useful for
the two functions to read data from sheets or CSVs.
}
\details{
This setup is not needed for working with folders/datafiles connected to your computer
via Google Drive Desktop (Mac/Windows), only for reading files directly
}
