% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gfile_modified_time}
\alias{gfile_modified_time}
\title{WIP get time stamp for a particular gfile}
\usage{
gfile_modified_time(gfile)
}
\arguments{
\item{gfile}{a file object from google drive}
}
\value{
timestamp value
}
\description{
WIP get time stamp for a particular gfile
}
