% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gsheet_auth_setup}
\alias{gsheet_auth_setup}
\title{setup authentication for reading google sheet}
\usage{
gsheet_auth_setup(drive_email)
}
\arguments{
\item{drive_email}{your preferred email, can be read from environment, set get_drive_email}
}
\value{
True/False if the authentication/setup was successful
}
\description{
this reads from the environment (or Renviron file) to get configuration details
for authenticating to a google sheets service
note that this is nearly identical to gdrive_setup but only for google sheets
google sheets has a different API and different permissions in the cloud console to read
}
