% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_commecologyrules_sheet}
\alias{read_commecologyrules_sheet}
\title{read in google sheet for formatted for specific project
data sheet features
\itemize{
\item has a 2nd row that is all text that needs to be removed
\item has the string "NA" in some cells that should be numeric only
\item since it's google some cols should be factors but that's not a thing google can do
\item date is not in standard scientific format
}}
\usage{
read_commecologyrules_sheet(gurl, sheet, spec.df)
}
\arguments{
\item{gurl}{google sheet url or ID per googlesheets package}

\item{sheet}{sheet tab name or number, forward to sheet param in  googlesheets4::read_sheet}

\item{spec}{data.frame that is the specification, must have columns col_name and col_type}
}
\value{
data.frame or NA if there is a problem
}
\description{
read in google sheet for formatted for specific project
data sheet features
\itemize{
\item has a 2nd row that is all text that needs to be removed
\item has the string "NA" in some cells that should be numeric only
\item since it's google some cols should be factors but that's not a thing google can do
\item date is not in standard scientific format
}
}
