% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{read_gsheet}
\alias{read_gsheet}
\title{read in google sheet into memory using our shared drive}
\usage{
read_gsheet(gsheet_name, shared_drive, verbose = "FALSE")
}
\arguments{
\item{gsheet_name}{required file name of the gsheet you need}
}
\value{
data frame as returned by read.csv, or NULL if there is an issue
}
\description{
reads the first sheet only, and the whole sheet.  Attemtps to load Google packages and log-in if necessary. In
the case there are two sheets with the same name, returns the first one it finds (not ideal! )
}
\details{
Note This is only for google sheets, not CSVs or other data files.  It assumes the 'share drive' is
is available in the environment
}
