% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{read_gsheet_by_url}
\alias{read_gsheet_by_url}
\title{read data in a google sheet given the sheets URL}
\usage{
read_gsheet_by_url(gurl, sheet_tab_number = 1, drive_email = NULL)
}
\description{
Note This is only for google sheets, not CSVs or other data files.
requires Oauth and google cloud console setup
}
