% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_url_list}
\alias{read_url_list}
\title{read in the table of google sheet URLs with id}
\usage{
read_url_list(gurl, id_column = "id", url_column = "url", drive_email = NULL)
}
\value{
dataframe with columns from google sheet with at least two columns 'id' and 'url' plus other columns in original sheet
}
\description{
google sheet requires two column on for id one for the url for that id
the names of which can be anything but
}
