% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{read_validate_and_save}
\alias{read_validate_and_save}
\title{given a URL and params, read, validate and save a CSV}
\usage{
read_validate_and_save(url, tab_name, spec.df, csv_folder = "../L0")
}
\description{
filename <- read_and_save(url, sheet_id = 'biomass_data', spec.df = commassembly_rules_biomass_str))
}
