% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{spec_to_readr_col_types}
\alias{spec_to_readr_col_types}
\title{convert our specification format to something useable by readr::read_csv}
\usage{
spec_to_readr_col_types(spec.df)
}
\arguments{
\item{spec.df}{dataframe with columns 'col_name' and 'data_str'}
}
\value{
list of col names and type abbreviations, per read_csv() docs
}
\description{
convert our specification format to something useable by readr::read_csv
}
