% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{type_code_to_readr_code}
\alias{type_code_to_readr_code}
\title{convert a type name to a readr convert code}
\usage{
type_code_to_readr_code(type_code)
}
\arguments{
\item{type_code}{single letter or string of type}
}
\value{
character single letter code
}
\description{
readr uses convert codes - see vignette("readr")
this function allows for named types and will convert those to the 1-letter
codes used by reader
}
