% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{type_converter_fun}
\alias{type_converter_fun}
\title{code to conversion function mapping}
\usage{
type_converter_fun(type_code)
}
\arguments{
\item{type_code}{character value indicating type, one of character, integer, factor, double, numeric, Date(capital D) or first letter}
}
\value{
one of the converter function as.integer... etc.  For Dates, return custom as.Date.flexible function defined above
}
\description{
given a character code, return the conversion function to use on a value.
This is useful for spreadsheets/csvs that don't adapt well to conversion (even with formats specific)
and hence manual converting from character is needed
Yes this is probably a re-make of what's in read.csv or read_csv but those weren't amenable to a specific
use case
}
