% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_all}
\alias{validate_all}
\title{read the list of urls, read, confirm column format, validate and
when using in development mode, make sure do run \code{devtools::load_all()}
to get all the commruleR functions loaded
save to CSV}
\usage{
validate_all(urls.df = NULL, drive_email = NULL)
}
\description{
read the list of urls, read, confirm column format, validate and
when using in development mode, make sure do run \code{devtools::load_all()}
to get all the commruleR functions loaded
save to CSV
}
