% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_biomass}
\alias{validate_biomass}
\title{run all validation checks for biomass}
\usage{
validate_biomass(biomass_df, validation_file = NULL)
}
\arguments{
\item{biomass_df}{data frame with biomass data}

\item{validation_file}{optional file to read for validation, will read from inst/rules if not sent}
}
\value{
boolean TRUE if all validation checks pass
}
\description{
run all validation checks for biomass
}
