% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_biomass_columns}
\alias{validate_biomass_columns}
\title{validate columns present for biomass data}
\usage{
validate_biomass_columns(biomass_data)
}
\description{
validate columns present for biomass data
}
