% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatR.R
\name{validate_data}
\alias{validate_data}
\title{validate data df}
\usage{
validate_data(data_df, spec_df, validation_rules)
}
\arguments{
\item{data_df}{dataframe of biomass data}

\item{spec_df}{data frame of table specification}

\item{validation_rules}{file with validation rules in it.}
}
\description{
use the validate package to check a file against a set of rules
}
