% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatR.R
\name{validate_data_columns}
\alias{validate_data_columns}
\title{validate columns against data definition}
\usage{
validate_data_columns(data_df, spec_df)
}
\arguments{
\item{data_df}{data frame following data definition}

\item{spec_df}{data frame of table specification}
}
\description{
validate columns against data definition
}
