% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_env_columns}
\alias{validate_env_columns}
\title{validate env columns}
\usage{
validate_env_columns(env_data)
}
\arguments{
\item{env_data}{data frame from imported env spreadsheet}
}
\description{
check that all columns present in imported env data
}
