% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatR.R
\name{validate_from_file}
\alias{validate_from_file}
\title{convenience to create validation results from a yaml file}
\usage{
validate_from_file(data_df, file)
}
\arguments{
\item{file}{yaml formatted file with rules for the validate package}

\item{data_dfe}{data frame of data}
}
\value{
the outupt from confront function
}
\description{
convenience to create validation results from a yaml file
}
