% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\docType{data}
\name{biomass_column_names}
\alias{biomass_column_names}
\title{constants for validation}
\format{
An object of class \code{character} of length 33.
}
\usage{
biomass_column_names
}
\description{
constants for validation
}
\keyword{datasets}
