% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{read_commrules_sheet}
\alias{read_commrules_sheet}
\title{read in one tab (only) of  meta-data analysis google sheet}
\usage{
read_commrules_sheet(sheet_url, has_comment_line = TRUE, drive_email = NULL)
}
\arguments{
\item{drive_email}{the email to use for google drive log-in, which is the institution that setup the project}

\item{sheeturl}{the url of of the google sheet - each tab has it's own URL!}

\item{remove_row_one}{datasheets use first row with explanatory text in google sheet, remove if TRUE.  defaults to TRUE}
}
\value{
data frame
}
\description{
read in one tab (only) of  meta-data analysis google sheet
}
