% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{read_gcsv}
\alias{read_gcsv}
\title{download a CSV from the project shared drive and read into memory}
\usage{
read_gcsv(file_name, shared_drive = NULL, file_path = NULL)
}
\arguments{
\item{shared_drive}{name of the shared drive to look in, defaults to the MSB project drive (2021)}

\item{filepath}{full name of the CSV file (e.g. myfile.csv ) with optional partial path.}
}
\value{
a data.frame as returned by read.csv, no row names.
}
\description{
given a CSV filename, find it in our shared drive and read it in.  If there are multiple files
found with the same name on the share drive, throws a warning and reads only the first one
it finds  (which may not be the most recent one!   )
}
\details{
This is not needed for working with folders/datafiles connected to your computer
via Google Drive Desktop (Mac/Windows), only for reading files directly.  If you don't have access
to the share drive it will not work
}
