% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_biomass_data}
\alias{validate_biomass_data}
\title{validate biomass data}
\usage{
validate_biomass_data(biomass_data, validation_file = biomass_validation_file)
}
\description{
use the validate package to check a file against the
}
